user.js changes
Removed tweaks that are no longer needed or are moving to in-browser settings options.

Web notifications and geolocation: moved to in-browser options, not a very big tracking vector anyways (require permission)

Browser/extension updates: on by default anyways

browser.pocket.enabled: no longer used

Enable telemetry: most people want to disable it, not enable it. Most of it's on by default anyways.