uMatrix rule addition
Fixed pocketcasts.com. Unfortunately, like OneDrive, there's a large list of domains that the website connects to, which changes based on who is accessing it and what they're accessing. It's basically unpredictable, so the only way to get the website to work is to allow all (that isn't on a blocklist).